pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[project]
name = "strands-agent-azure-openai-demos"
version = "0.1.0"
description = "Examples of using Strands Agents SDK with OpenAI and Azure OpenAI"
authors = [{ name = "Henry Skiba", email = "henry.skiba@frgrisk.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["strands-agents-tools (>=0.2.2,<0.3.0)", "strands-agents[litellm] (>=1.1.0,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
